The IzPack packaging of Silverpeas 6.1 was full of bugs: - the Silverpeas construction in SilverpeasPreparation.groovy wasn't correct as   it was performed in another directory that the one from which the Silverpeas   installation will be done (two SILVERPEAS_HOME!) - some times the copy or the move of the resources into the   target/classes/package weren't correctly done (depends on the rights settings   of the files to copy or to move) - a bug in the name of the SQL script for H2
These are now fixed.

Because H2 doesn't support the windowing function COUNT() OVER() that is used
in Silverpeas in the pagination of data, it is necessary to disable the use
of this windowing function by setting the property
jdbc.pagination.method.countOver at false in the properties file
properties/org/silverpeas/general.properties.